US-11181-11787 [PROGRESS] OUT: Added warning message when partially processing and OUT with a signature. Fixed state in signature for PDF

+++ bin/addons/msf_outgoing/wizard/internal_move_processor.py
@@ -86,6 +86,7 @@ class internal_picking_processor(osv.osv):
             readonly=True,
         ),
         'draft': fields.boolean('Draft'),
+        'partial_process_sign': fields.boolean('Partial process warning if signature'),
     }
 
     def default_get(self, cr, uid, fields_list=None, context=None, from_web=False):
@@ -328,12 +329,16 @@ class internal_picking_processor(osv.osv):
                 _('No data to process !'),
             )
 
+        res = {}
+        partial_and_signed = False
         to_unlink = []
 
         for proc in wizards:
             total_qty = 0.00
 
+            l_ids = []
             for line in proc.move_ids:
+                l_ids.append(line.id)
                 # if no quantity, don't process the move
                 if not line.quantity:
                     to_unlink.append(line.id)
@@ -360,11 +365,37 @@ class internal_picking_processor(osv.osv):
                     _('You have to enter the quantities you want to process before processing the move.'),
                 )
 
-        # Remove non-used lines
-        if to_unlink:
+            # Add the warning if there's a signed signature during partial processing
+            if self._name == 'outgoing.delivery.processor' and l_ids and proc.picking_id.signature_id \
+                    and not proc.partial_process_sign:
+                has_signed = False
+                for sign_line in proc.picking_id.signature_id.signature_line_ids:
+                    if sign_line.signed:
+                        partial_and_signed = True
+                        has_signed = True
+                        break
+                if has_signed:
+                    cr.execute("""SELECT id FROM outgoing_delivery_move_processor WHERE id IN %s 
+                            AND quantity < ordered_quantity LIMIT 1""", (tuple(l_ids),))
+                    if cr.fetchone():
+                        self.write(cr, uid, proc.id, {'partial_process_sign': True}, context=context)
+                        res = {
+                            'type': 'ir.actions.act_window',
+                            'res_model': self._name,
+                            'res_id': proc.id,
+                            'view_type': 'form',
+                            'view_mode': 'form',
+                            'target': 'new',
+                            'view_id': [self.pool.get('ir.model.data').get_object_reference(cr, uid, 'msf_outgoing',
+                                                                                            'outgoing_delivery_processor_form_view')[1]],
+                            'context': context,
+                        }
+
+        # Remove non-used lines if the lines are not partially processed when the document is signed
+        if to_unlink and not partial_and_signed:
             proc_line_obj.unlink(cr, uid, to_unlink, context=context)
 
-        return True
+        return res
 
     def do_reset(self, cr, uid, ids, context=None):
         if context is None:
